refactor(runtime): deslop subagent runtime tests
