+++ packages/runtime/src/session/session-tool-loop.test.ts
@@ -1,0 +1,46 @@
+import type { ModelMessage } from "ai";
+import { describe, expect, it } from "vitest";
+import { Agent } from "../agent";
+import {
+  assistantMessage,
+  toolCallPart,
+  toolResultFor,
+  userText,
+} from "../test-fixtures";
+import { userTextToModelMessage } from "./mapping";
+import { collect } from "./session.test-support";
+
+describe("Agent session tool loop", () => {
+  it("continues the model loop after a tool call result", async () => {
+    const seenHistory: ModelMessage[][] = [];
+    let calls = 0;
+    const agent = new Agent({
+      llm: ({ history }) => {
+        calls += 1;
+        seenHistory.push([...history]);
+
+        if (calls === 1) {
+          const toolCall = toolCallPart("call-tool-loop-1");
+          return Promise.resolve([
+            assistantMessage([toolCall]),
+            toolResultFor(toolCall),
+          ]);
+        }
+
+        return Promise.resolve([assistantMessage("DONE")]);
+      },
+    });
+
+    await collect(await agent.send("remember me"));
+
+    const toolCall = toolCallPart("call-tool-loop-1");
+    expect(seenHistory).toEqual([
+      [userTextToModelMessage(userText("remember me"))],
+      [
+        userTextToModelMessage(userText("remember me")),
+        assistantMessage([toolCall]),
+        toolResultFor(toolCall),
+      ],
+    ]);
+  });
+});
